+++ on-fe/app/(auth)/login/additional_info/info-form.css
@@ -49,10 +49,10 @@ line-height: normal;
 /* 레이아웃 */
 display: flex;
 width: 357px;
-height: 74px;
+height: 10dvw;
 padding: 1px 5px;
 align-items: flex-start;
-gap: 10px;
+/* gap: 10px; */
 flex-shrink: 0;
 
 /* 스타일 */
@@ -60,6 +60,8 @@ border: none; /* 전체 테두리 제거 */
 outline: none; /* 포커스 시 파란 테두리 제거 (선택사항) */
 border-bottom: 1px solid var(--LightGray, #BEBEBE);
 background: var(--White, #FFF);
+
+font-size: 20px;
 }
 
 .title{
