+++ on-fe/app/(auth)/login/additional_info/page.js
@@ -1,38 +1,15 @@
 'use client';
+import AdditionalInfoForm from '@/components/auth/AdditionalInfoForm';
 import './info-form.css';
+import FormLayout from '@/components/form/FormLayOut';
 
 export default function AdditionalInfoPage() {
     return (
-        <div className='bg'>
-        <div className="additional-info">
-            <div className='inputs'>
-
-                <p className="logo">ON</p>
-
-                <div>
-                    <label className='title'>*이름</label>
-                    <input type="text" className='input'></input>
-                </div>
-                <div>
-                    <label className='title'>*닉네임</label>
-                    <input type="text" className='input'></input>
-                </div>
-                <div>
-                    <label className='title'>*생년월일 (8자리) </label>
-                    <input type="text" maxLength="8" className='input'></input>
-                </div>
-
-                <div className='checkbox-form'>
-                    <input type="checkbox" style={{ accentColor: '#ffd753', marginRight: '8px' }} />
-                    <span>이용약관 개인정보 수집 및 이용, 마케팅 활용 선택에 모두 동의합니다.</span>
-                </div>
-                
-
-                <button className="submit-btn">회원가입</button>
-
-            </div>
-            
-        </div>
-        </div>
+        // <div className='bg'>
+        //     <AdditionalInfoForm/>
+        // </div>
+        <FormLayout>
+            <AdditionalInfoForm/>
+        </FormLayout>
     );
 }
