+++ on-fe/app/globals.css
@@ -1,5 +1,5 @@
 @import url('https://cdn.jsdelivr.net/npm/pretendard@latest/dist/web/static/pretendard.css');
-/* @import "tailwindcss"; */
+@import "tailwindcss";
 
 * {
   font-family: 'Pretendard';
@@ -8,4 +8,5 @@
 html, body {
   width: 100%;
   height: 100%;
+  overflow: hidden;
 }
